Add initial C++ test cases
